creating a new package for just the lens interface defined in lens.actor
