+++ conftest.py
@@ -1,1 +1,1 @@
-collect_ignore = ['setup.py']
+collect_ignore = ['setup.py', 'setup-interface.py']
