+++ lens/environment/boot.py
@@ -1,9 +1,19 @@
 '''
-boot functions for initialize environment simulations.
+boot functions for initialize environment simulations:
+Can be called from the command line
+by specifying an environment "agent_type" and an outer-id:
 
-Can be called from the command line by specifying an environment "agent_type" and an outer-id:
+>> python -m lens.environment.boot --type lattice --id lattice
 
->>python -m lens.environment.boot --type lattice --id lattice
+boot functions for compartment simulations:
+Can be called from the command line
+after booting an environmental "outer" simulation and passing in the outer-id, the agent_type:
+
+>> python -m lens.environment.boot --outer-id 'outer-id' --type 'agent_type'
+
+Once an environment and agent are booted, they can be triggered with:
+
+>> python -m lens.environment.control run --id 'outer-id'
 
 '''
 
@@ -12,11 +22,55 @@ from __future__ import absolute_import, division, print_function
 import os
 import shutil
 
+from lens.actor.inner import Inner
 from lens.actor.outer import Outer
 from lens.actor.boot import BootAgent
 from lens.actor.emitter import get_emitter
+
 from lens.environment.lattice import EnvironmentSpatialLattice
 from lens.environment.make_media import Media
+from lens.environment.lattice_compartment import generate_lattice_compartment
+
+# processes
+from lens.processes.transport_lookup import TransportLookup
+from lens.processes.CovertPalsson2002_metabolism import Metabolism
+from lens.processes.Kremling2007_transport import Transport
+from lens.processes.growth import Growth
+from lens.processes.Endres2006_chemoreceptor import ReceptorCluster
+
+# composites
+from lens.composites.Covert2008 import initialize_covert2008
+from lens.composites.growth_division import initialize_growth_division
+
+def wrap_boot(initialize, initial_state):
+    def boot(agent_id, agent_type, agent_config):
+        initial_state.update(agent_config.get('declare', {}))
+        agent_config['declare'] = initial_state  # 'declare' is for the environment
+
+        return Inner(
+            agent_id,
+            agent_type,
+            agent_config,
+            initialize)
+
+    return boot
+
+def wrap_initialize(make_process):
+    def initialize(boot_config):
+        boot_config.update({
+            'emitter': {
+                'type': 'database',
+                'url': 'localhost:27017',
+                'database': 'simulations',
+                },
+            'compartment_options':{
+                'time_step': 1.0,
+                },
+            })
+        process = make_process(boot_config)  # 'boot_config', set in environment.control is the process' initial_parameters
+        return generate_lattice_compartment(process, boot_config)
+
+    return initialize
 
 
 DEFAULT_COLOR = [0.6, 0.4, 0.3]
@@ -255,6 +309,16 @@ class BootEnvironment(BootAgent):
             'sugar1_small': boot_glc_g6p_small,
             'sugar2': boot_glc_lct,
             'measp': boot_measp,
+
+            # single process compartments
+            'lookup': wrap_boot(wrap_initialize(TransportLookup), {'volume': 1.0}),
+            'metabolism': wrap_boot(wrap_initialize(Metabolism), {'volume': 1.0}),
+            'transport': wrap_boot(wrap_initialize(Transport), {'volume': 1.0}),
+            'growth': wrap_boot(wrap_initialize(Growth), {'volume': 1.0}),
+            'receptor': wrap_boot(wrap_initialize(ReceptorCluster), {'volume': 1.0}),
+            # composite compartments
+            'growth_division': wrap_boot(initialize_growth_division, {'volume': 1.0}),
+            'covert2008': wrap_boot(initialize_covert2008, {'volume': 1.0})
             }
 
 def run():
