+++ requirements-interface.txt
@@ -1,0 +1,4 @@
+confluent-kafka==0.11.5
+numpy==1.14.6
+pymongo==3.6.1
+pytest==4.6.5
